+++ src/main/java/com/wallmart/exceptions/ProductNotFoundException.java
@@ -1,7 +1,7 @@
 package com.wallmart.exceptions;
 
 public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(Long itemId) {
+    public ProductNotFoundException(int itemId) {
         super("Product not found with ID: " + itemId);
     }
 }
